Introduce option to check frame bound consistency
This restores the behaviour of previous version to NOT check it. This is
not valid according to strict reading of the specification but it may
appear in practice. In particular, imagemagick handles this in weird
ways that are not errors but silently adjust the width and height.. Not
sure about that style.