Fix DecodeSubBlock bytes returned, add allow_unknown blocks decoding option
`StreamingDecoder::next_state` could return a byte offset that was
larger than the current buffer. This could result in a panic on the
next loop iteration while indexing into the buffer with the offset.

Add an option to DecodeOptions for allowing the processing of unknown
blocks in `BlockStart`. Previously, all unknown blocks would return an
error if the `BlockStart` was not `Image`, `Extension`, or `Trailer`.