Fix infinite loop during decoding
In frames with empty image data it was possible to trigger a condition
where decoding would be stuck, neither reading any part of the data nor
producing any LZW decoded data. The outer loop expects to make progress
and odes not have any alternative stop condition. This was an infinite
blocking loop.