Better handle reset codes in LZW stream
Backport of commit: 3912dbf2b613d048787c43cd9fec1d0b6ad884ee

The status condition from `weezl` was not being interpreted correctly.
Instead of looking at the status to decide whether decoding was stalled,
the decision was based on the consumed and produce byte counts. When
handling reset codes that are buffered internally in the LZW decoder,
neither is advanced.

Avoid this bug by forwarding the status condition to the block stage of
the decoder where we can continue with further state machine actions as
long as the decoder makes any progress. (We still must be prepared for
an image not containing enough data). The decoder itself already
diagnoses the condition as `NoProgress`.